src/requirements.txt
requests==2.28.2
openai==0.25.0
pydantic==1.10.2
pyhumps==3.8.0
boto3==1.26.39